+++ src/RayTracing/ColorCalculator.java
@@ -30,14 +30,14 @@ public class ColorCalculator {
         MyVector directionToLight = new MyVector(intersection.position, light.position);
         MyVector normal = intersection.surface.get_normal(intersection.position);
         MyVector reflectionDirection =(directionToLight.multiply(2).projectTo(normal)).subtract(directionToLight);
-        float intensity = light.specularIntensity*
-                (float) Math.pow(reflectionDirection.getAbsCosAngel(intersection.direction), intersection.surface.material.phong_coefficient);
+        double intensity = light.specularIntensity*
+                (double) Math.pow(reflectionDirection.getAbsCosAngel(intersection.direction), intersection.surface.material.phong_coefficient);
         return intersection.surface.material.specular_color.multiply(light.color).multiply(intensity);
     }
 
     private Color diffuseColor(Light light){
         MyVector directionToLight = new MyVector(intersection.position, light.position);
-        float intensity = intersection.surface.get_normal(intersection.position).getAbsCosAngel(directionToLight);
+        double intensity = intersection.surface.get_normal(intersection.position).getAbsCosAngel(directionToLight);
         return light.color.multiply(intersection.surface.material.defuse_color).multiply(intensity);
     }
