+++ src/RayTracing/Light.java
@@ -3,11 +3,11 @@ package RayTracing;
 public class Light {
     public MyVector position;
     public Color color;
-    public float specularIntensity;
-    public float shadow;
-    public float radius;
+    public double specularIntensity;
+    public double shadow;
+    public double radius;
 
-    public Light(MyVector position, Color color, float specular, float shadow, float radius) {
+    public Light(MyVector position, Color color, double specular, double shadow, double radius) {
         this.position = position;
         this.color = color;
         this.specularIntensity = specular;
